# docs/swagger.yaml
basePath: /api
definitions:
...
  model.AdminCreateLinkRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      logo:
        type: string
      name:
        type: string
      siteshot:
        type: string
      skip_health_check:
        type: boolean
      sort_order:
        type: integer
      status:
        enum:
        - PENDING
        - APPROVED
        - REJECTED
        - INVALID
        type: string
      tag_id:
        description: 改为单个标签，可选
        type: integer
      url:
        type: string
    required:
    - category_id
    - name
    - status
    - url
    type: object
...
  model.LinkCategoryDTO:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      style:
        type: string
    type: object
  model.LinkDTO:
    properties:
      category:
        $ref: '#/definitions/model.LinkCategoryDTO'
      description:
        type: string
      id:
        type: integer
      logo:
        type: string
      name:
        type: string
      siteshot:
        type: string
      skip_health_check:
        type: boolean
      sort_order:
        type: integer
      status:
        type: string
      tag:
        allOf:
        - $ref: '#/definitions/model.LinkTagDTO'
        description: 改为单个标签
      url:
        type: string
    type: object
...
  model.LinkListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/model.LinkDTO'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  model.LinkTagDTO:
    properties:
      color:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  model.ReviewLinkRequest:
    properties:
      siteshot:
        type: string
      status:
        enum:
        - APPROVED
        - REJECTED
        type: string
    required:
    - status
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@anheyu.com
    name: 安知鱼
    url: https://github.com/anzhiyu-c/anheyu-app
  description: Anheyu App 应用接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Anheyu App API
  version: "1.0"
paths:
...
  /links:
    get:
      description: 获取所有友链列表，支持分页和筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 审核状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.LinkListResponse'
              type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取友链列表（管理员）
      tags:
      - 友链管理
    post:
      consumes:
      - application/json
      description: 管理员手动创建新友链
      parameters:
      - description: 友链信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AdminCreateLinkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.LinkDTO'
              type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 创建失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建友链
      tags:
      - 友链管理
  /links/{id}:
    delete:
      description: 管理员删除指定友链
      parameters:
      - description: 友链ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 删除失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除友链
      tags:
      - 友链管理
    put:
      consumes:
      - application/json
      description: 管理员更新友链信息
      parameters:
      - description: 友链ID
        in: path
        name: id
        required: true
        type: string
      - description: 友链信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AdminUpdateLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.LinkDTO'
              type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新友链
      tags:
      - 友链管理
  /links/{id}/review:
    put:
      consumes:
      - application/json
      description: 管理员审核友链申请（批准/拒绝）
      parameters:
      - description: 友链ID
        in: path
        name: id
        required: true
        type: string
      - description: 审核信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.ReviewLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 审核成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 审核失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 审核友链
      tags:
      - 友链管理
  /links/categories:
    get:
      description: 获取所有友链分类列表，包括统计信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.LinkCategoryDTO'
                  type: array
              type: object
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取分类列表（管理员）
      tags:
      - 友链管理
    post:
      consumes:
      - application/json
      description: 创建友链分类
      parameters:
      - description: 分类信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateLinkCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.LinkCategoryDTO'
              type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 创建失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建分类
      tags:
      - 友链管理
  /links/categories/{id}:
    delete:
      description: 删除友链分类
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数无效或删除失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除分类
      tags:
      - 友链管理
    put:
      consumes:
      - application/json
      description: 更新友链分类信息
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: string
      - description: 分类信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateLinkCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.LinkCategoryDTO'
              type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新分类
      tags:
      - 友链管理
...
  /links/import:
    post:
      consumes:
      - application/json
      description: 批量导入友链数据（最多1000个）
      parameters:
      - description: 导入的友链数据
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.ImportLinksRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 导入完成
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.ImportLinksResponse'
              type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 导入失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 批量导入友链
      tags:
      - 友链管理
  /links/sort:
    put:
      consumes:
      - application/json
      description: 批量更新友链的显示顺序
      parameters:
      - description: 排序信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.BatchUpdateLinkSortRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 批量更新友链排序
      tags:
      - 友链管理
  /links/tags:
    get:
      description: 获取所有友链标签列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.LinkTagDTO'
                  type: array
              type: object
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取标签列表
      tags:
      - 友链管理
    post:
      consumes:
      - application/json
      description: 创建友链标签
      parameters:
      - description: 标签信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateLinkTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.LinkTagDTO'
              type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 创建失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建标签
      tags:
      - 友链管理
  /links/tags/{id}:
    delete:
      description: 删除友链标签
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数无效或删除失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除标签
      tags:
      - 友链管理
    put:
      consumes:
      - application/json
      description: 更新友链标签信息
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: string
      - description: 标签信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateLinkTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.LinkTagDTO'
              type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新标签
      tags:
      - 友链管理
...
  /public/link-categories:
    get:
      description: 获取包含已批准友链的分类列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.LinkCategoryDTO'
                  type: array
              type: object
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取公开分类列表
      tags:
      - 友情链接
...
  /public/links/random:
    get:
      description: 随机获取指定数量的已批准友链
      parameters:
      - default: 0
        description: 获取数量，0表示全部
        in: query
        name: num
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.LinkDTO'
                  type: array
              type: object
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      summary: 随机获取友链
      tags:
      - 友情链接
...
securityDefinitions:
  BearerAuth:
    description: '在请求头中添加 Bearer Token，格式为: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"